requirements.txt
psycopg2-binary==2.9.9 # Only way around pg config error
logging
